VeSyncSuperior6000S: remove automatic_stop methods
VeSyncSuperior6000S does not support automatic_off (automatically
turning off when humidity is with 5% of the targeet). This is not the
same as auto mist mode.

The automatic_stop_on/automatic_stop_off methods were incorrectly
setting the humidity (as if copied from set_auto_mode/set_manual_mode).

I opted to remove the method instead of returning False which does not
indicate if the operation is supported. Python docs suggest doing this
instead of raising NotImplementedError.
https://docs.python.org/3.13/library/exceptions.html#NotImplementedError